requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Configuration
django-environ==0.4.0
django-secure==1.0.1

# Forms
django-braces==1.8.1
django-crispy-forms==1.6.0
django-floppyforms==1.6.2

# Models
django-model-utils==2.5

# Images
Pillow==3.2.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.25.2

# Python-MariaDB and Python-MySQL Database Adapter
mysqlclient==1.3.7

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.4

# Redis support
django-redis==4.4.3
redis>=2.10.0

# Your custom requirements go here
django-tinycontent==0.5.1
django-bleach==0.3.0
django-teryt-tree==0.6.0
django-versatileimagefield==1.4
https://github.com/ad-m/django-atom/archive/v0.11.0.zip
django_filter==0.13
django-autocomplete-light==3.1.6
django-tinymce==2.3.0
cached_property==1.3.0
django-extra-views==0.7.1
djangorestframework==3.3.3
Markdown==2.6.6
jsonfield==1.0.3
django-bootstrap-pagination==1.6.1
# Raven is the Sentry client
# --------------------------
raven

